feat: replace SectionHeader component with a shared version, adding memoization and optional description prop
